+++ e2e/wdio/platform/tests/select.e2e-spec.ts
@@ -3,6 +3,7 @@ import {
     click,
     getAttributeByName,
     getElementArrayLength,
+    getElementSize,
     getText,
     isElementClickable,
     refreshPage,
@@ -24,6 +25,7 @@ describe('Select test suite', () => {
     const {
         selectModeExample,
         displayText,
+        select,
         buttons,
         options,
         selectedValue_1,
@@ -94,6 +96,14 @@ describe('Select test suite', () => {
         it('should be able to select the option', () => {
             checkOptions(customControlContentExample, 48);
         });
+
+        it('should check changing width of select after selection', () => {
+            const defaultSelectWidth = getElementSize(customControlContentExample + select, 0, 'width');
+
+            checkOptions(customControlContentExample, 48);
+            const newSelectWidth = getElementSize(customControlContentExample + select, 0, 'width');
+            expect(newSelectWidth).toBeGreaterThan(defaultSelectWidth);
+        });
     });
 
     describe('Check Select In Mobile Mode example', () => {
@@ -153,7 +163,7 @@ describe('Select test suite', () => {
         });
     });
 
-    function checkOptions(selector: string, id: number, index: number = 0): void {
+    function checkOptions(selector: string, id: number): void {
         const textBefore = getText(selector + displayText);
         click(selector + buttons);
         click(options(id));