+++ libs/platform/src/lib/form/text-area/text-area.component.ts
@@ -245,6 +245,7 @@ export class TextAreaComponent extends BaseInput implements AfterViewChecked, On
     /** update the counter message and related interactions */
     updateCounterInteractions(): void {
         this._textAreaCharCount = this.value?.length ?? 0;
+
         if (this.maxLength) {
             // newly added to avoid unnecessary iteration, remove if issue found
             this.validateLengthOnCustomSet();
@@ -258,29 +259,23 @@ export class TextAreaComponent extends BaseInput implements AfterViewChecked, On
                 this._targetElement.focus();
                 this._targetElement.setSelectionRange(this.maxLength, this._textAreaCharCount);
             }
+
             this.counterExcessOrRemaining = this.excessText;
             this.exceededCharCount = this._textAreaCharCount - this.maxLength;
         } else {
             this.counterExcessOrRemaining = this.remainingText;
             this.exceededCharCount = this.maxLength - this._textAreaCharCount;
         }
-        // reset flag
+
         this._isPasted = false;
     }
 
     /** handle exceeded maxlength case when text is pasted */
     handlePasteInteraction(): void {
         this._isPasted = true;
-        /// For IE
-        if (window['clipboardData']) {
-            // todo: handle for IE
-            this.updateCounterInteractions();
-        } else {
-            // for other navigators
-            navigator['clipboard'].readText().then(() => {
-                this.updateCounterInteractions();
-            });
-        }
+
+        // Wrapped in timeout to make selecting range working
+        setTimeout(() => this.updateCounterInteractions());
     }
 
     /** handle auto growing of textarea */